Day 10 Making right about my assignment

+++ addBlockToBoard.c
@@ -3,6 +3,8 @@
 int main(void) {
     RemoveCursor();
     speed = 10;
+    curPosX = (GBOARD_WIDTH + GBOARD_ORIGIN_X) / 2;
+    curPosY = GBOARD_ORIGIN_Y;
     srand((unsigned int)time(NULL));
     block_id = (rand() % 7) * 4;
 
@@ -31,10 +33,10 @@ void AddBlockToBoard(void) {
 
 int DetectCollision(int posX, int posY, char blockModel[4][4]) {
     int x, y;
-    int arrX = posX - GBOARD_ORIGIN_X;
+    int arrX = (posX - GBOARD_ORIGIN_X) / 2;
     int arrY = posY - GBOARD_ORIGIN_Y;
 
-    for (y = 0; y < 4; x++) {
+    for (y = 0; y < 4; y++) {
         for (x = 0; x < 4; x++) {
             if ((gameBoardInfo[arrY + y][arrX + x] == 1) && (blockModel[y][x] == 1))
                 return (0);
@@ -104,8 +106,8 @@ COORD GetCurrentCursorPos(void) {
 }
 
 void ProcessKeyInput(void) {
-    //if (!DetectCollision(curPosX, curPosY + 1, blockModel[block_id]))
-    //    return;
+    if (!DetectCollision(curPosX, curPosY + 1, blockModel[block_id]))
+        return;
 
     int key;
 
@@ -140,18 +142,18 @@ void ShiftRight(void) {
 
 void ShiftLeft(void) {
     if (!DetectCollision(curPosX - 2, curPosY, blockModel[block_id])) {
-        return;
+       return;
     }
     DeleteBlock(blockModel[block_id]);
-    curPosY -= 2;
+    curPosX -= 2;
     SetCurrentCursorPos(curPosX, curPosY);
     ShowBlock(blockModel[block_id]);
 }
 
 int BlockDown(void) {
-    //if (!DetectCollision(curPosX, curPosY + 1, blockModel[block_id])) {
-    //  return (0);
-    //}
+    if (!DetectCollision(curPosX, curPosY + 1, blockModel[block_id])) {
+      return (0);
+    }
     DeleteBlock(blockModel[block_id]);
     curPosY += 1;
     SetCurrentCursorPos(curPosX, curPosY);
@@ -167,7 +169,7 @@ void ReverseRotateBlock(void) {
         return;
     }
     DeleteBlock(blockModel[block_id]);
-    int block_id = block_rotated;
+    block_id = block_rotated;
     
     ShowBlock(blockModel[block_id]);
 }
@@ -195,15 +197,16 @@ void DeleteBlock(char blockInfo[4][4]) {
 
 void ShowBlock(char blockInfo[4][4]) {
     int y, x;
-    
+    COORD curPos = GetCurrentCursorPos();
+
     for (y = 0; y < 4; y++) {
         for (x = 0; x < 4; x++) {
-            SetCurrentCursorPos(curPosX + (x * 2), curPosY + y);
+            SetCurrentCursorPos(curPos.X + (x * 2), curPos.Y + y);
 
             if (blockInfo[y][x] == 1) {
                 printf("бс");
             }
         }
     }
-    SetCurrentCursorPos(curPosX, curPosY);
+    SetCurrentCursorPos(curPos.X, curPos.Y);
 }
